Remove call to accordion.destroy.
When working with Backbone regions, I'm not entirely sure when/if to handle
jQuery-UI widget destruction.  It seems that replacing the view in a
Backbone region would teardown any DOM elements, thereby destroying the
jQuery-UI widget.  And I haven't been able to *see* any leakage.  (But maybe
I don't know how to look.) In any event, we actually have other jQuery-UI
widgets on the page which are part of regions, and the accordion is the only
one for which we were doing any cleanup.  For consistency, we'll remove that
cleanup call.